+++ HousingSearchListener/Infrastructure/BaseService.cs
@@ -1,18 +1,15 @@
 ﻿using System;
 using System.Diagnostics.CodeAnalysis;
 using Amazon.XRay.Recorder.Handlers.AwsSdk;
-using Elasticsearch.Net;
 using Hackney.Core.Logging;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
-using Nest;
 
 namespace HousingSearchListener.Infrastructure
 {
     [ExcludeFromCodeCoverage]
-    public abstract class BaseFunction
+    public abstract class BaseService
     {
         protected IConfigurationRoot Configuration { get; }
 
@@ -20,7 +17,7 @@ namespace HousingSearchListener.Infrastructure
 
         protected ILogger Logger { get; }
 
-        protected BaseFunction(IServiceCollection services)
+        protected BaseService(IServiceCollection services)
         {
             AWSSDKHandler.RegisterXRayForAllServices();
 
@@ -34,7 +31,7 @@ namespace HousingSearchListener.Infrastructure
             ServiceProvider = services.BuildServiceProvider();
             ServiceProvider.UseLogCall();
 
-            Logger = ServiceProvider.GetRequiredService<ILogger<BaseFunction>>();
+            Logger = ServiceProvider.GetRequiredService<ILogger<BaseService>>();
         }
 
         private void AddLogging(IServiceCollection services)
@@ -51,21 +48,6 @@ namespace HousingSearchListener.Infrastructure
         protected virtual void ConfigureServices(IServiceCollection services)
         {
             services.AddLogCallAspect();
-        }
-    }
-
-    public static class ESServiceInitialization
-    {
-        public static void ConfigureElasticsearch(IServiceCollection services)
-        {
-            var url = Environment.GetEnvironmentVariable("ELASTICSEARCH_DOMAIN_URL") ?? "http://localhost:9200";
-            var pool = new SingleNodeConnectionPool(new Uri(url));
-            var connectionSettings =
-                new ConnectionSettings(pool)
-                    .PrettyJson().ThrowExceptions().DisableDirectStreaming();
-            var esClient = new ElasticClient(connectionSettings);
-
-            services.TryAddSingleton<IElasticClient>(esClient);
         }
     }
 }
